refactor(suite): repeated wallet header refactored
