+++ packages/suite/src/components/wallet/index.tsx
@@ -2,6 +2,7 @@ import AccountsMenu from './AccountsMenu';
 import DiscoveryProgress from './DiscoveryProgress';
 import Title from './Title';
 import WalletLayout from './WalletLayout';
+import WalletLayoutHeader from './WalletLayout/components/WalletLayoutHeader';
 import CoinBalance from './CoinBalance';
 import OnOffSwitcher from './OnOffSwitcher';
 import InputError from './InputError';
@@ -27,6 +28,7 @@ export {
     Title,
     AccountsMenu,
     WalletLayout,
+    WalletLayoutHeader,
     DiscoveryProgress,
     CoinBalance,
     CoinmarketLayout,
