+++ packages/suite/src/views/wallet/send/components/Header/index.tsx
@@ -1,36 +1,11 @@
 import React from 'react';
-import styled from 'styled-components';
-import { variables, Dropdown, H2 } from '@trezor/components';
-import { Translation, AccountFormCloseButton } from '@suite-components';
+import { Dropdown } from '@trezor/components';
+import { Translation } from '@suite-components';
 import { useActions } from '@suite-hooks';
 import { useSendFormContext } from '@wallet-hooks';
 import * as sendFormActions from '@wallet-actions/sendFormActions';
 import Clear from './components/Clear';
-
-const Wrapper = styled.div`
-    display: flex;
-    margin-bottom: 30px;
-`;
-
-const HeaderLeft = styled.div`
-    display: flex;
-    flex: 1;
-    align-items: center;
-`;
-
-const HeaderRight = styled.div`
-    display: flex;
-    align-items: center;
-    justify-content: flex-end;
-    flex: 1;
-`;
-
-const StyledTitle = styled(H2)`
-    font-weight: ${variables.FONT_WEIGHT.DEMI_BOLD};
-    text-transform: capitalize;
-    color: ${props => props.theme.TYPE_DARK_GREY};
-    padding-top: 10px;
-`;
+import { WalletLayoutHeader } from '@wallet-components';
 
 const Header = () => {
     const {
@@ -74,27 +49,19 @@ const Header = () => {
     ];
 
     return (
-        <Wrapper>
-            <HeaderLeft>
-                <StyledTitle>
-                    <Translation id="TR_NAV_SEND" />
-                </StyledTitle>
-            </HeaderLeft>
-            <HeaderRight>
-                <Clear />
-                <Dropdown
-                    alignMenu="right"
-                    data-test="@send/header-dropdown"
-                    items={[
-                        {
-                            key: 'header',
-                            options,
-                        },
-                    ]}
-                />
-                <AccountFormCloseButton />
-            </HeaderRight>
-        </Wrapper>
+        <WalletLayoutHeader title="TR_NAV_SEND">
+            <Clear />
+            <Dropdown
+                alignMenu="right"
+                data-test="@send/header-dropdown"
+                items={[
+                    {
+                        key: 'header',
+                        options,
+                    },
+                ]}
+            />
+        </WalletLayoutHeader>
     );
 };
